SAF USB Use Common utility
:Release Notes:
SAF USB Use Common utility

:Detailed Notes:
SAF USB Use Common utility

:Testing Performed:
Build Successful & Verified SAF Functionality

:QA Notes:
N/A

:Issues Addressed:
[PLAT-143061 "Use Common Utility for USB file
              operations"

Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.storageaccess/+/311218
Tested-by: <phanendra.koppisetti@lgepartner.com>
Tested-by: <mukul20.jha@lge.com>
Tested-by: <navnit.kumar@lge.com>

+++ src/utils/SAFUtilityOperation.h
@@ -40,12 +40,12 @@ private:
     std::string mPath;
     std::string mType;
     std::string mModTime;
-    uint32_t mSize;
+    uintmax_t mSize;
 
     void init();
     std::string getFileType(std::string);
     std::string getLastWrite(std::string);
-    uint32_t getFileSize(std::string);
+    uintmax_t getFileSize(std::string);
     std::string getModTime();
 
 public:
@@ -53,7 +53,7 @@ public:
     std::string getName() { return mName; }
     std::string getPath() { return mPath; }
     std::string getType() { return mType; }
-    uint32_t getSize() { return mSize; }
+    uintmax_t getSize() { return mSize; }
     std::string getLastModTime() { return mModTime; }
 };
